D5: decompose economy_cog into cogs/economy/ + views/economy/
Reduces cogs/economy_cog.py from 1154 → 334 LOC (~71% reduction)
by hoisting the shared data tables (JOBS, SHOP_ITEMS, daily tiers,
cooldown constants) and helper functions into a subpackage, and by
moving all view classes — including the PersistentView panel — into
views/economy/.

New layout:
  disbot/cogs/economy/
  ├── __init__.py
  └── _helpers.py        JOBS / SHOP_ITEMS / _DAILY_TIERS /
                         _WORK_COOLDOWN / _DAILY_COOLDOWN +
                         _daily_weights / _pick_daily / _job_pay /
                         _available_jobs / _shop_embed /
                         _build_economy_embed

  disbot/views/economy/
  ├── __init__.py
  ├── main_panel.py      EconomyPanelView (PersistentView, SUBSYSTEM="economy")
  ├── work_panel.py      _WorkView / _JobSelect / _WorkSubView
  └── shop_panel.py      _ShopView / _ShopSelect / _ShopSubView /
                         _ShopPanelSelect

Persistence-rule check:
- EconomyPanelView keeps SUBSYSTEM = "economy" unchanged.
- Every button's custom_id (economy:daily / work / shop / balance /
  inventory / jobs / overview) is preserved verbatim.
- panel_anchors and runtime_sessions rows continue to resolve.
- The @register decorator on EconomyPanelView fires when
  views.economy is imported by cogs/economy_cog.py at module load,
  so persistent_views._REGISTRY entry is established before on_ready
  calls restore_anchors().

Back-compat re-exports: cogs/economy_cog.py re-imports every
previously module-level public name (JOBS, SHOP_ITEMS, _pick_daily,
EconomyPanelView, _WorkView, …) so external imports keep resolving.
The _WorkSubView and _ShopSubView "Back" buttons use late imports to
break the views ↔ views circular reference cleanly.

Verification: ruff/black/isort/mypy clean; pytest 436/436 green.
